requirements.txt
requests==2.31.0
supabase==2.3.4
python-dotenv==1.0.0
 
